feat: C,U,D endpoints for article require authorization as admin

+++ Backend/Controllers/ArticleController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Backend.Interfaces;
 using Backend.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,9 +13,11 @@ namespace Backend.Controllers {
   [ApiController]
   public class ArticleController : ControllerBase {
     private readonly IArticleRepository _articleRepository;
+    private readonly ILoginRepository _loginRepository;
 
-    public ArticleController(IArticleRepository articleRepository) {
+    public ArticleController(IArticleRepository articleRepository, ILoginRepository loginRepository) {
       _articleRepository = articleRepository;
+      _loginRepository = loginRepository;
     }
 
     // GET: api/articles
@@ -33,13 +36,21 @@ namespace Backend.Controllers {
 
     // POST: api/article
     [HttpPost]
+    [Authorize(AuthenticationSchemes = "Bearer")]
     public IActionResult Post([FromBody] Article article) {
+      var usr = _loginRepository.GetUserFromToken(Request.Headers["Authorization"]);
+      if (!usr.is_admin) return Unauthorized();
+
       return Ok(_articleRepository.Create(article));
     }
 
     // PATCH: api/article
     [HttpPatch("{id}")]
+    [Authorize(AuthenticationSchemes = "Bearer")]
     public IActionResult Patch(int id, [FromBody] Article article) {
+      var usr = _loginRepository.GetUserFromToken(Request.Headers["Authorization"]);
+      if (!usr.is_admin) return Unauthorized();
+
       Article a = _articleRepository.Patch(id, article.title, article.text, article.image_url);
       if (a == null) return NotFound();
       return Ok(a);
@@ -47,7 +58,11 @@ namespace Backend.Controllers {
 
     // DELETE: api/article
     [HttpDelete("{id}")]
+    [Authorize(AuthenticationSchemes = "Bearer")]
     public IActionResult Delete(int id) {
+      var usr = _loginRepository.GetUserFromToken(Request.Headers["Authorization"]);
+      if (!usr.is_admin) return Unauthorized();
+
       try {
         _articleRepository.Delete(id);
       }
